+++ __tests__/controllers/engines/engine.view.test.js
@@ -1,5 +1,6 @@
-const { getCompilationConfigs, getBlueprint, getRuntimeMethods, getTemplate, buildTemplate, getFileString, render,
-  Template } = require('../../../controllers/engines/view')[Symbol.for('UNIT_TEST')];
+const {
+  getCompilationConfigs, getBlueprint, getRuntimeMethods, getTemplate, buildTemplate, getFileString, render, Template,
+} = require('../../../controllers/engines/view')[Symbol.for('UNIT_TEST')];
 
 
 // mock