+++ __tests__/controllers/utilities/object.test.js
@@ -1,88 +1,96 @@
-// module
-const { checkToStringTag, hasOwnKey, cloneDeep, mergeDeep, assignDeep, freezeDeep,
-  proxyfyInCaseInsensitiveKey } = require('../../../controllers/utilities/')[Symbol.for('UNIT_TEST')];
+const {
+  hasOwnKey, cloneDeep, mergeDeep, assignDeep, freezeDeep, proxyfyInCaseInsensitiveKey,
+} = require('../../../controllers/utilities/')[Symbol.for('UNIT_TEST')];
 
 
 // test
-describe('Check the ENV', () => {
-  test('Should run in test mode', () => {
-    expect(process.env.NODE_ENV).toEqual('test');
-  });
-});
-
-
-describe('Bundle: Object methods', () => {
-  test('Fn: checkToStringTag: check the native brand(type) of objects', () => {
-    expect(checkToStringTag(async () => {})).toBe('AsyncFunction');
-    expect(checkToStringTag(() => {}, 'AsyncFunction')).toBeFalsy();
-    expect(checkToStringTag([() => {}], 'array')).toBeTruthy();
-  });
-
-  test('Fn: hasOwnKey: check if object has wwn a property', () => {
-    expect(hasOwnKey({}, '')).toBeFalsy();
-    expect(hasOwnKey({ a: 0 }, 'a')).toBeTruthy();
-    expect(hasOwnKey({ a: { b: 0 } }, 'b')).toBeFalsy();
-  });
-
-  test('Fn: cloneDeep: clone the object/array deeply by its value (reference decoupled)', () => {
-    const mockSource1 = { test: 'test' };
-    const mockSource2 = ['test'];
-    const mockSource3 = { a: { b: Symbol('test') } };
-    const result = [cloneDeep(mockSource1), cloneDeep(mockSource2), cloneDeep(mockSource3)];
+describe('Utilities: Object', () => {
+  test('Fn: hasOwnKey', () => {
+    const target = [
+      {},
+      { a: 0 },
+      { b: { a: 0 } },
+    ];
     //
-    expect(result[0]).not.toBe(mockSource1);
-    expect(result[0]).toEqual(mockSource1);
-    expect(result[1]).not.toBe(mockSource2);
-    expect(result[1]).toEqual(mockSource2);
-    expect(result[2]).not.toBe(mockSource3);
-    expect(result[2]).toEqual(mockSource3);
-    expect(result[2].a.b).toBe(mockSource3.a.b);
-    expect(result[2].a.b).not.toBe(Symbol('test'));
+    const test = target.map(obj => hasOwnKey(obj, 'a'));
+    expect(test[0]).toBeFalsy();
+    expect(test[1]).toBeTruthy();
+    expect(test[2]).toBeFalsy();
   });
 
-  test('Fn: mergeDeep: merge two object recursively', () => {
-    const mockTarget = { a: 0, b: 1, c: { d: { e: 2, f: 3 }, g: 4 } };
-    const mockSource = { c: { d: { e: 5 }, f: 6 } };
-    const result = [mergeDeep(mockTarget, mockSource), mergeDeep(mockSource, mockTarget, { mutate: true })];
+
+  test('Fn: cloneDeep', () => {
+    const target = [
+      ['a'],
+      { a: 'a' },
+      { b: [{ a: 'a' }] },
+    ];
     //
-    expect(result[0]).not.toEqual(mockTarget);
-    expect(result[0]).toEqual({ a: 0, b: 1, c: { d: { e: 5, f: 3 }, f: 6, g: 4 } });
-    expect(result[1]).toBe(mockSource);
-    expect(result[1]).toEqual({ a: 0, b: 1, c: { d: { e: 2, f: 3 }, f: 6, g: 4 } });
+    const test = target.map(obj => cloneDeep(obj));
+    expect(test[0]).toEqual(target[0]);
+    expect(test[0]).not.toBe(target[0]);
+    expect(test[1]).toEqual(target[1]);
+    expect(test[1]).not.toBe(target[1]);
+    expect(test[2]).toEqual(target[2]);
+    expect(test[2]).not.toBe(target[2]);
   });
 
-  test('Fn: freezeDeep: frozen the target and its property deeply', () => {
-    const mockTarget1 = { a: { b: { c: { d: { e: { f: 0 } } } } } };
-    const mockTarget2 = { a: { b: { c: { d: { e: { f: 0 } } } } } };
-    const result = [freezeDeep(mockTarget1), freezeDeep(mockTarget2, { mutate: true })];
+
+  test('Fn: mergeDeep', () => {
+    const target = { a: 0, b: 1, c: { d: { e: 2, f: 3 }, g: 4 } };
+    const source = { c: { d: { e: 5 }, f: 6 } };
+    const result = { a: 0, b: 1, c: { d: { e: 5, f: 3 }, f: 6, g: 4 } };
     //
-    expect(result[0]).not.toBe(mockTarget1);
-    expect(result[1]).toBe(mockTarget2);
-    expect(Object.getOwnPropertyDescriptors(mockTarget1)).toHaveProperty('a.writable', true);
-    expect(Object.getOwnPropertyDescriptors(mockTarget2)).toHaveProperty('a.writable', false);
-    expect(Object.getOwnPropertyDescriptors(result[0])).toHaveProperty('a.writable', false);
-    expect(Object.getOwnPropertyDescriptors(result[1])).toHaveProperty('a.writable', false);
-    expect(Object.getOwnPropertyDescriptors(result[0].a.b.c.d.e)).toHaveProperty('f.writable', false);
-    expect(Object.getOwnPropertyDescriptors(result[1].a.b.c.d.e)).toHaveProperty('f.writable', false);
+    const test_1 = mergeDeep(target, source);
+    expect(test_1).toEqual(result);
+    expect(target).not.toEqual(result);
+    //
+    const test_2 = mergeDeep(target, source, { mutate: true });
+    expect(test_2).toEqual(result);
+    expect(target).toEqual(result);
   });
 
-  test('Fn: assignDeep: assigned the value to an object by path recursively', () => {
-    const mockTarget1 = {};
-    const mockTarget2 = {};
-    const mockPath = 'a[b].c.d[e][f]';
-    const result = [assignDeep(mockTarget1, mockPath, 0), assignDeep(mockTarget2, mockPath, 0, { mutate: true })];
+
+  test('Fn: freezeDeep', () => {
+    const target = { a: { b: { c: { d: { e: { f: 0 } } } } } };
+    expect(Object.getOwnPropertyDescriptors(target)).toHaveProperty('a.writable', true);
+    expect(Object.getOwnPropertyDescriptors(target.a.b.c.d.e)).toHaveProperty('f.writable', true);
     //
-    expect(result[0]).not.toBe(result[1]);
-    expect(result[0]).toEqual({ a: { b: { c: { d: { e: { f: 0 } } } } } });
-    expect(result[1]).toEqual({ a: { b: { c: { d: { e: { f: 0 } } } } } });
+    const test_1 = freezeDeep(target);
+    expect(test_1).not.toBe(target);
+    expect(Object.getOwnPropertyDescriptors(test_1)).toHaveProperty('a.writable', false);
+    expect(Object.getOwnPropertyDescriptors(test_1.a.b.c.d.e)).toHaveProperty('f.writable', false);
+    //
+    const test_2 = freezeDeep(target, { mutate: true });
+    expect(test_2).toBe(target);
+    expect(Object.getOwnPropertyDescriptors(test_2)).toHaveProperty('a.writable', false);
+    expect(Object.getOwnPropertyDescriptors(test_2.a.b.c.d.e)).toHaveProperty('f.writable', false);
   });
 
-  test('Fn: proxyfyInCaseInsensitiveKey: proxyfy the object for allowing case-insensitive access', () => {
-    const mockTarget = { a: 1 };
-    const result = proxyfyInCaseInsensitiveKey(mockTarget);
+
+  test('Fn: assignDeep', () => {
+    const target = { a: { b: { g: 'g' } } };
+    const path = 'a[b].c.d[e][f]';
+    const value = 'f';
+    const result = { a: { b: { c: { d: { e: { f: 'f' } } }, g: 'g' } } };
     //
-    expect(mockTarget.A).toBeUndefined();
-    expect(result.A).toEqual(result.a);
-    expect(result.A).toBe(1);
+    const test_1 = assignDeep(target, path, value);
+    expect(test_1).not.toBe(target);
+    expect(test_1).toEqual(result);
+    //
+    const test_2 = assignDeep(target, path, value, { mutate: true });
+    expect(test_2).toBe(target);
+    expect(test_2).toEqual(result);
+  });
+
+
+  test('Fn: proxyfyInCaseInsensitiveKey', () => {
+    const target = { a: 1 };
+    const test = proxyfyInCaseInsensitiveKey(target);
+    //
+    expect(test).not.toBe(target);
+    expect(test).toEqual(target);
+    expect(target.A).toBeUndefined();
+    expect(test.A).toEqual(test.a);
   });
 });